third_party/WebKit/LayoutTests/fast/html/imports/import-custom-element-onload.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
hello = Object.create(HTMLElement.prototype);
hello.ids = [];
hello.createdCallback = function() {
    hello.ids.push(this.id);
};

document.registerElement('x-hello', {prototype: hello});

t = async_test('Custom elements in an import are upgraded inside load event handlers of the import.');

function ready() {
    t.step(function() {
        assert_array_equals(['1', '2', '3'], hello.ids);
        t.done();
    });
}
</script>
<link rel="import" href="resources/import-custom-element-onload-child.html" onload="ready()">
</head>
<body>
</body>
</html>
